Support Realm (0.93.2)
Abolished the XCTest of encryption.
See Also: https://github.com/realm/realm-cocoa/issues/1885

+++ Classes/YSRealmStore/YSRealmStore.h
@@ -23,8 +23,8 @@ typedef void(^YSRealmStoreWriteTransactionCompletion)(YSRealmStore *store, YSRea
 
 @optional
 - (void)migrationWithMigration:(RLMMigration *)migration
-              oldSchemaVersion:(NSUInteger) oldSchemaVersion;
-- (NSUInteger)schemaVersion;
+              oldSchemaVersion:(uint64_t) oldSchemaVersion;
+- (uint64_t)schemaVersion;
 
 @end
